Made the following changes:   - Tagged everything with OLD_KEYBOARD_EVENTS.   - Overhauled keyboard events. The event data is now stored via     iEvent's property bag mechanism (a helper class can simplify     the access.) The idea of 'scan codes' and 'char codes' has been     replaced with 'raw codes' and 'cooked codes.' Also, there's only     one keyboard event now, and the information whether the key was     pressed or released is now part of the event data. A couple of     features have been added as well, for example, CS can now     distinguish between left/right modifiers or keypad and     non-keypad keys.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15690 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/perftest/ptests2d.cpp
@@ -87,7 +87,7 @@ void StringTester::Setup (iGraphics3D* g3d, PerfTest* /*perftest*/)
   Tester2D::Setup (g3d, 0);
   font = G2D->GetFontServer ()->LoadFont (CSFONT_SMALL);
   char tl [] = "aA1! bB2 cC3 dD4$ 5%eE fF^6 gG&7hH8*iI9(j0)Jk[K1lL]}mM@n2No+O5p> Qq<r;RSs Tt =Uu|uv;Vwxyz WXYZ Crystal Space, the open sourced graphics engine.";
-  int tl_len = strlen (tl);
+  size_t tl_len = strlen (tl);
   int line_length;
   font->GetDimensions (tl, line_length, text_height);
   float ave_char_w = ((float)line_length)/tl_len;
